+++ domain/src/main/java/com/mashup/gabbangzip/sharedalbum/domain/model/LoginParam.kt
@@ -1,0 +1,8 @@
+package com.mashup.gabbangzip.sharedalbum.domain.model
+
+data class LoginParam(
+    val idToken: String,
+    val provider: String = "KAKAO",
+    val nickname: String,
+    val profileImage: String,
+)
