+++ domain/src/main/java/com/mashup/gabbangzip/sharedalbum/domain/usecase/LoginUseCase.kt
@@ -1,5 +1,6 @@
 package com.mashup.gabbangzip.sharedalbum.domain.usecase
 
+import com.mashup.gabbangzip.sharedalbum.domain.model.LoginParam
 import com.mashup.gabbangzip.sharedalbum.domain.repository.LoginRepository
 import javax.inject.Inject
 
@@ -10,10 +11,3 @@ class LoginUseCase @Inject constructor(
         return runCatching { loginRepository.login(param) }
     }
 }
-
-data class LoginParam(
-    val idToken: String,
-    val provider: String = "KAKAO",
-    val nickname: String,
-    val profileImage: String,
-)
